aider: Added observer for `AcaiaScaleWeight` notification in `viewDidLoad()` and implemented `onWeightUpdate(_:)` method to update dose with weight from the scale.
# Context:
USER: have a look at `AiderTest/AiderTest/ViewController.swift` for how to handle weight updates from the scale
ASSISTANT: Looking at the `AiderTest/AiderTest/ViewController.swift` file, I see that the weight updates from the scale are handled in the `onWeightUpdate(_:)` method. This method is called whenever the `AcaiaScaleWeight` notification is posted.

To update the `Dose` variable in `PreShotDataViewController` with the current weight from the scale, we need to observe the `AcaiaScaleWeight` notification in `PreShotDataViewController` and call the `updateDose(with:)` method in the notification handler.

Here are the changes:

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    override func viewDidLoad() {
        super.viewDidLoad()
        // Setup UI elements here
        setupUI()
    }
=======
    override func viewDidLoad() {
        super.viewDidLoad()
        // Setup UI elements here
        setupUI()
        NotificationCenter.default.addObserver(self, selector: #selector(onWeightUpdate(_:)), name: NSNotification.Name(rawValue: AcaiaScaleWeight), object: nil)
    }
>>>>>>> UPDATED
```

In the updated code, I've added an observer for the `AcaiaScaleWeight` notification in the `viewDidLoad()` method. This will call the `onWeightUpdate(_:)` method whenever the `AcaiaScaleWeight` notification is posted.

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    @objc func doseButtonPressed() {
        doseSet = true
    }

    // This method should be called when new data is received from the scale
    func updateDose(with weight: Double) {
        if !doseSet {
            Dose = weight
        }
    }
=======
    @objc func doseButtonPressed() {
        doseSet = true
    }

    @objc func onWeightUpdate(_ notification: NSNotification) {
        guard let weight = notification.userInfo?[AcaiaScaleUserInfoKeyWeight] as? Double else { return }
        updateDose(with: weight)
    }

    // This method should be called when new data is received from the scale
    func updateDose(with weight: Double) {
        if !doseSet {
            Dose = weight
        }
    }
>>>>>>> UPDATED
```

In the updated code, I've added the `onWeightUpdate(_:)` method. This method gets the weight from the notification's `userInfo` dictionary and calls the `updateDose(with:)` method with the weight.