aider: Updated `Grind` and `RPM` variables to be computed properties that update the text of `grindLabel` and `rpmLabel` respectively.
# Context:
USER: can we make sure the `grindLabel` and `rpmLabel` represent the values in the `Grind` and `RPM` variable as they change?
ASSISTANT: Sure, we can achieve this by updating the `Grind` and `RPM` variables to be computed properties. This way, whenever their values are updated, we can also update the text of the `grindLabel` and `rpmLabel` respectively.

Here are the changes:

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    var Grind: Double = 17.0
    var RPM: Double = 100.0
=======
    var Grind: Double = 17.0 {
        didSet {
            grindLabel.text = "Grind: \(Grind)"
        }
    }
    var RPM: Double = 100.0 {
        didSet {
            rpmLabel.text = "RPM: \(RPM)"
        }
    }
>>>>>>> UPDATED
```

This code will update the `grindLabel` and `rpmLabel` text whenever the `Grind` and `RPM` variables are updated.