aider: Added a new variable `updateDose` and set it to `true`, and updated the `doseButtonPressed()` and `updateDose(with weight: Double)` functions to use this variable to control the updating of the `Dose` variable.
# Context:
USER: in `AiderTest/AiderTest/PreShotDataViewController.swift` can we use the set button to stop updating the value of the dose variable when it is pressed by using a vairiable "updateDose = false"
ASSISTANT: To implement this change, we need to modify the `AiderTest/AiderTest/PreShotDataViewController.swift` file.

We will add a new variable `updateDose` and set it to `true` initially. When the `doseButton` is pressed, we will set `updateDose` to `false`. In the `updateDose(with weight: Double)` function, we will only update the `Dose` if `updateDose` is `true`.

Here are the changes:

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    var doseSet: Bool = false
    var Grind: Double = 17.0 {
=======
    var doseSet: Bool = false
    var updateDose: Bool = true
    var Grind: Double = 17.0 {
>>>>>>> UPDATED
```

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    @objc func doseButtonPressed() {
        doseSet = true
    }
=======
    @objc func doseButtonPressed() {
        doseSet = true
        updateDose = false
    }
>>>>>>> UPDATED
```

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    func updateDose(with weight: Double) {
        if !doseSet {
            Dose = weight
        }
    }
=======
    func updateDose(with weight: Double) {
        if !doseSet && updateDose {
            Dose = weight
        }
    }
>>>>>>> UPDATED
```