aider: Updated the `onWeightUpdate` function to only update the `Dose` variable if `updateDose` is true.
# Context:
USER: my `doseLabel` is still updating with the scale weigth even when we've pressed the set button]
ASSISTANT: To stop updating the `doseLabel` when the set button is pressed, we need to modify the `onWeightUpdate(_ notification: NSNotification)` function in the `AiderTest/AiderTest/PreShotDataViewController.swift` file.

We will only update the `Dose` variable if `updateDose` is `true`.

Here is the change:

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    @objc func onWeightUpdate(_ notification: NSNotification) {
        guard let weight = notification.userInfo?[AcaiaScaleUserInfoKeyWeight] as? Double else { return }
        let truncatedWeight = round(weight * 1000) / 1000
        Dose = truncatedWeight
        updateDose(with: truncatedWeight)
    }
=======
    @objc func onWeightUpdate(_ notification: NSNotification) {
        guard let weight = notification.userInfo?[AcaiaScaleUserInfoKeyWeight] as? Double else { return }
        let truncatedWeight = round(weight * 1000) / 1000
        if updateDose {
            Dose = truncatedWeight
        }
        updateDose(with: truncatedWeight)
    }
>>>>>>> UPDATED
```